add all file

+++ BusinessObject/Work_on.h
@@ -10,14 +10,14 @@ class Work_on : public BusinessObject {
 private:
     static int IdNew;   //    Identification of work_on sheet for update  
     int Id;             //    The current identification of work_on sheet
-    double Essn;       //    Social security number of employee 
+    int Essn;       //    Social security number of employee 
     int Pno;           //    Number of project 
-    float Hours;      //    Overtime hours  
+    double Hours;       //    Overtime hours  
 public:
     //Default constructor
     Work_on(); 
     //Contructor with input parameters     
-    Work_on(double Essn, int Pno, double Hours); 
+    Work_on(int Essn, int Pno, double Hours); 
     void IncreaseId();
     //GetId() is a function that get ID value of work_on sheet
     int& GetId();
